Improved error messages in views.py
When handling error messages returned from the MWS server, we use the
json package to pretty-print the error message to stderr. However, the
message didn't come from MWS (for instances a 404 error handled by
Tomcat before it made it to MWS), then the error message won't be JSON
formatted and trying to pretty-print it will cause an unhandled
exception.  This change handles that exception.