Fix ACL's on user-generated files
Files created by the user's job (as opposed to uploaded from MantidPlot)
need to have their ACL's set so that apache can read them.  Originally,
we did this by queuing up a separate script to call setfacl using the
atd daemon.  However, atd on fermi seems to be broken and thus the ACL's
weren't being set.  This change removes the need for using atd.  The
script to set the ACL's is called using setsid instead.